gem loading benchmark needs to have the in-fork require fixed also
